Refactor proxy protos, lru cache for factories
